meesho_oms/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
psycopg2-binary==2.9.9
Pillow==10.3.0
pytesseract==0.3.10
pdf2image==1.17.0
python-dotenv==1.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
openpyxl==3.1.2
hashlib2==1.0.0
Werkzeug==3.0.3
gunicorn==22.0.0
